Use df.to_dict('records') instead of iterrows
Replaces the use of `iterrows` in notebook 5, with `df.to_dict('records')`
which creates a list of dictionary, one for each row. As a result, adding the
trip counts is about 10 times faster.

This also removes the `set_index` argument to `read_csv`, otherwise you'd need
to do a `reset_index()` before `to_dict` because it only includes columns.